Clears library contents and re-renders library
when adding a new book instead of inserting in to the library array,
creating a new book, creating a new card for that book, and finally
appending it to the library node, just clear out the whole library
and re-render it. This makes it easier to use data index to delete
from the library array as we won't have to update each card's data index
after doing so.